refactor: reorganize quiz types and interfaces into separate file
- Moved quiz-related enums and interfaces from
- dashboard.ts to a new quiz.ts file for better modularity and maintainability.
- Updated dashboard.ts to re-export commonly used quiz types for convenience.
